backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pyodbc==5.2.0
bcrypt==4.3.0
python-jose==3.5.0
pydantic[email]==2.11.7
python-dotenv==1.1.1
